Fix backend + refactor store names
