+++ frontend/src/app/app.config.ts
@@ -3,8 +3,8 @@ import {provideRouter} from '@angular/router';
 import {routes} from './app.routes';
 import {providePrimeNG} from 'primeng/config';
 import LaraLightBlue from '@primeuix/themes/aura';
-import {petReducer} from './store/pet-store/pet.reducer';
-import {PetEffects} from './store/pet-store/pet.effects';
+import {petReducer} from './store/pet/pet.reducer';
+import {PetEffects} from './store/pet/pet.effects';
 import {authReducer} from './store/auth/auth.reducer';
 import {AuthEffects} from './store/auth/auth.effects';
 import {provideStore} from '@ngrx/store';
