+++ frontend/src/app/component/pet-list/pet-list-component.html
@@ -68,7 +68,7 @@
   <button pButton type="button" class="reset-search-button" (click)="onReset($event)"> Reset</button>
 </form>
 
-<p-dataView #dv [value]="(pets$ | async ) || []" [rows]="3" [paginator]="true">
+<p-dataView #dv [value]="(pets$ | async ) || []" [rows]="10" [paginator]="true">
   <ng-template #list let-items>
     <div class="items-panel">
       @for (pet of items; track pet.id) {
@@ -84,7 +84,7 @@
             </div>
           </div>
           <div class="buttons-panel">
-            <p-button styleClass="w-full md:w-auto">Details</p-button>
+            <p-button styleClass="w-full md:w-auto" [routerLink]="['/pet-profile', pet.id]">Details</p-button>
           </div>
         </div>
       }
