+++ frontend/src/app/component/pet-list/pet-list-component.ts
@@ -2,8 +2,8 @@ import {Component, OnDestroy, OnInit} from '@angular/core';
 import {DataViewModule} from 'primeng/dataview';
 import {Button, ButtonDirective} from 'primeng/button';
 import {Store} from '@ngrx/store';
-import {loadPets, searchPets} from '../../store/pet-store/pet.actions';
-import {selectAllPets, selectPetStatus, selectPetError} from '../../store/pet-store/pet.selectors';
+import {loadPets, searchPets} from '../../store/pet/pet.actions';
+import {selectAllPets, selectPetStatus, selectPetError} from '../../store/pet/pet.selectors';
 import {PetDTO} from '../../api/model/petDTO';
 import {Observable, startWith, Subject, takeUntil, map, switchMap} from 'rxjs';
 import {AsyncPipe} from '@angular/common';
