+++ frontend/src/app/component/pet-profile/pet-profile.ts
@@ -1,15 +1,17 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { AsyncPipe, NgClass } from '@angular/common';
 import { Store } from '@ngrx/store';
-import { Observable } from 'rxjs';import { PetDTO } from '../../api';
-import { loadPet, adoptPet } from '../../store/pet.actions';
-import { selectSelectedPet, selectPetStatus, selectPetError } from '../../store/pet.selectors';
+import { Observable } from 'rxjs';
+import { PetDTO } from '../../api';
+import { loadPet, adoptPet } from '../../store/pet/pet.actions';
+import { selectSelectedPet, selectPetStatus, selectPetError } from '../../store/pet/pet.selectors';
 import { ActivatedRoute } from '@angular/router';
+import {NavBar} from '../nav-bar/nav-bar';
 
 @Component({
   selector: 'app-pet-profile',
   standalone: true,
-  imports: [AsyncPipe, NgClass],
+  imports: [AsyncPipe, NgClass, NavBar],
   templateUrl: './pet-profile.html',
   styleUrls: ['./pet-profile.css']
 })
