+++ frontend/src/app/store/pet/pet.actions.ts
@@ -1,5 +1,5 @@
 import { createAction, props } from '@ngrx/store';
-import { PetDTO } from '../../api/model/petDTO';
+import { PetDTO } from '../../api';
 
 export const loadPets = createAction('[Pet List] Load Pets');
 export const loadPetsSuccess = createAction('[Pet List] Load Pets Success', props<{ pets: PetDTO[] }>());
